Lets the user decide how big the font size should be for an article.
Added an item to the menu article_view.xml so the user can select the font
size.  Shows a dialog box with three options:  small, medium, and large.
Also I deleted the search option in the article_view.xml because the user
doesn't need that in an article view.  Strings.xml was changed to
accommodate new strings.